requirements.txt
alembic==1.13.1
APScheduler==3.10.4
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
flask==3.0.2
Flask-APScheduler==1.13.1
Flask-Migrate==4.0.7
flask-sqlalchemy==3.1.1
greenlet==3.0.3
idna==3.6
importlib-metadata==7.0.2
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
pycparser==2.21
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
schedule==1.2.1
six==1.16.0
SQLAlchemy==2.0.28
typing-extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
werkzeug==3.0.1
zipp==3.18.1
